[clang-tidy] Enable readability-redundant-casting
Drop explicit casts that were converting an expression to its own type.
Found via:
  script/clang-tidy --fix --environment esp32-arduino-tidy --all-headers

after temporarily removing -readability-redundant-casting from
.clang-tidy. clang-format was then run over the modified files.

clang-tidy is AST-aware and only flagged casts where the sub-expression
already has the target type — e.g. uint16_t(uint16_t_val) — while
leaving real type conversions (uint8_t -> uint32_t, etc.) intact.

48 sites across 30 files. No semantic change.

Tracked in esphome/backlog#114.

+++ esphome/components/ac_dimmer/ac_dimmer.cpp
@@ -133,7 +133,7 @@ void IRAM_ATTR HOT AcDimmerDataStore::gpio_intr() {
       if (this->method == DIM_METHOD_LEADING_PULSE) {
         // Minimum pulse time should be enough for the triac to trigger when it is close to the ZC zone
         // this is for brightness near 99%
-        this->disable_time_us = std::max(this->enable_time_us + GATE_ENABLE_TIME, (uint32_t) cycle_time_us / 10);
+        this->disable_time_us = std::max(this->enable_time_us + GATE_ENABLE_TIME, cycle_time_us / 10);
       } else {
         this->gate_pin.digital_write(false);
         this->disable_time_us = this->cycle_time_us;
